src/data_ingestion/web_scraper.py
# src/data_ingestion/web_scraper.py
import requests
from bs4 import BeautifulSoup
from typing import List, Dict, Any
import time
import logging
from config import Config

logger = logging.getLogger(__name__)

class LegalWebScraper:
    def __init__(self):
        """Initialize web scraper for legal documents"""
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
        self.sources = Config.LEGAL_SOURCES
    
    def scrape_constitution_updates(self) -> List[Dict[str, Any]]:
        """Scrape latest constitutional updates"""
        updates = []
        
        try:
            # This is a placeholder - you'll need to implement specific scrapers
            # for each government website based on their structure
            
            # Example for a generic government site
            url = self.sources.get('constitution', '')
            if url:
                response = requests.get(url, headers=self.headers, timeout=10)
                soup = BeautifulSoup(response.content, 'html.parser')
                
                # Extract text content (customize based on site structure)
                content_divs = soup.find_all('div', class_='content')
                
                for div in content_divs:
                    text_content = div.get_text(strip=True)
                    if len(text_content) > 100:  # Filter out short content
                        updates.append({
                            'content': text_content,
                            'source': 'Constitution Updates',
                            'url': url,
                            'scraped_date': time.strftime('%Y-%m-%d')
                        })
            
            logger.info(f"Scraped {len(updates)} constitutional updates")
            return updates
            
        except Exception as e:
            logger.error(f"Error scraping constitutional updates: {e}")
            return []
    
    def scrape_recent_judgments(self) -> List[Dict[str, Any]]:
        """Scrape recent Supreme Court judgments"""
        judgments = []
        
        try:
            # Placeholder for Supreme Court website scraping
            # Note: Government websites often have specific terms of use
            # Make sure to comply with their robots.txt and terms
            
            url = self.sources.get('supreme_court', '')
            # Implementation would go here based on the actual website structure
            
            logger.info(f"Scraped {len(judgments)} recent judgments")
            return judgments
            
        except Exception as e:
            logger.error(f"Error scraping recent judgments: {e}")
            return []
    
    def scrape_new_legislation(self) -> List[Dict[str, Any]]:
        """Scrape new legislation and amendments"""
        legislation = []
        
        try:
            # Placeholder for legislative updates
            # Implementation would depend on the specific government portals
            
            logger.info(f"Scraped {len(legislation)} legislative updates")
            return legislation
            
        except Exception as e:
            logger.error(f"Error scraping new legislation: {e}")
            return []
    
    def get_all_updates(self) -> List[Dict[str, Any]]:
        """Get all legal updates from various sources"""
        all_updates = []
        
        # Collect updates from all sources
        all_updates.extend(self.scrape_constitution_updates())
        all_updates.extend(self.scrape_recent_judgments())
        all_updates.extend(self.scrape_new_legislation())
        
        # Add rate limiting to be respectful to government servers
        time.sleep(2)
        
        logger.info(f"Total scraped updates: {len(all_updates)}")
        return all_updates
