requirements.txt
# requirements.txt
streamlit==1.28.1
chromadb==0.4.15
sentence-transformers==2.2.2
transformers==4.35.0
torch
langchain==0.0.335
langchain-community==0.0.5
python-docx==0.8.11
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow
requests==2.31.0
beautifulsoup4==4.12.2
pandas
numpy
python-dotenv
fastapi
uvicorn